+++ src/oss.c
@@ -89,7 +89,7 @@ static int ossinit(sox_format_t * ft)
 
     if (ft->signal.channels > 2) ft->signal.channels = 2;
 
-    if (ioctl(fileno(ft->fp), SNDCTL_DSP_RESET, 0) < 0)
+    if (ioctl(fileno(ft->fp), (size_t) SNDCTL_DSP_RESET, 0) < 0)
     {
         lsx_fail_errno(ft,SOX_EOF,"Unable to reset OSS driver.  Possibly accessing an invalid file/device");
         return(SOX_EOF);
@@ -188,14 +188,14 @@ static int ossinit(sox_format_t * ft)
     file->size = 0;
     ioctl (fileno(ft->fp), SNDCTL_DSP_GETBLKSIZE, &file->size);
     if (file->size < 4 || file->size > 65536) {
-            lsx_fail_errno(ft,SOX_EOF,"Invalid audio buffer size %d", file->size);
+            lsx_fail_errno(ft,SOX_EOF,"Invalid audio buffer size %lu", (unsigned long)file->size);
             return (SOX_EOF);
     }
     file->count = 0;
     file->pos = 0;
     file->buf = lsx_malloc(file->size);
 
-    if (ioctl(fileno(ft->fp), SNDCTL_DSP_SYNC, NULL) < 0) {
+    if (ioctl(fileno(ft->fp), (size_t) SNDCTL_DSP_SYNC, NULL) < 0) {
         lsx_fail_errno(ft,SOX_EOF,"Unable to sync dsp");
         return (SOX_EOF);
     }
